+++ common/src/main/java/net/mehvahdjukaar/moonlight/api/client/texture_renderer/RenderedTexturesManager.java
@@ -11,16 +11,14 @@ import com.mojang.blaze3d.vertex.*;
 import net.mehvahdjukaar.moonlight.api.util.Utils;
 import net.mehvahdjukaar.moonlight.core.Moonlight;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.gui.screens.inventory.MerchantScreen;
+import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.client.renderer.entity.ItemRenderer;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.block.Blocks;
 import org.jetbrains.annotations.Nullable;
 import org.joml.Matrix4f;
-import org.lwjgl.glfw.GLFW;
-import org.lwjgl.opengl.GL11;
 
 import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
@@ -76,12 +74,6 @@ public class RenderedTexturesManager {
         return texture;
     }
 
-
-    @Deprecated(forRemoval = true)
-    public static FrameBufferBackedDynamicTexture getFlatItemStackTexture(ResourceLocation res, ItemStack stack, int size) {
-        return requestFlatItemStackTexture(res, stack, size);
-    }
-
     public static FrameBufferBackedDynamicTexture requestFlatItemStackTexture(ResourceLocation res, ItemStack stack, int size) {
         return requestTexture(res, size, t -> drawItem(t, stack), true);
     }
@@ -129,7 +121,6 @@ public class RenderedTexturesManager {
         drawAsInGUI(tex, s -> {
             //render stuff
             ItemRenderer itemRenderer = Minecraft.getInstance().getItemRenderer();
-            //Minecraft.getInstance().gui.render(s,1);
             itemRenderer.renderGuiItem(s, stack, 0, 0);
         });
     }
@@ -137,10 +128,11 @@ public class RenderedTexturesManager {
     /**
      * Coordinates here are from 0 to 1
      */
-    public static void drawAsInWorld(FrameBufferBackedDynamicTexture tex, Consumer<PoseStack> drawFunction) {
+    public static void drawNormalized(FrameBufferBackedDynamicTexture tex, Consumer<PoseStack> drawFunction) {
         drawAsInGUI(tex, s -> {
             float scale = 1f / 16f;
-            s.scale(scale, scale, scale);
+            s.translate(8,8,0);
+            s.scale(scale, scale, 1);
             drawFunction.accept(s);
         });
     }
@@ -150,19 +142,18 @@ public class RenderedTexturesManager {
      * If you render an item at 0,0 it will be centered
      */
     public static void drawAsInGUI(FrameBufferBackedDynamicTexture tex, Consumer<PoseStack> drawFunction) {
-
         Minecraft mc = Minecraft.getInstance();
         RenderTarget frameBuffer = tex.getFrameBuffer();
         frameBuffer.clear(Minecraft.ON_OSX);
+
         //render to this one
         frameBuffer.bindWrite(true);
 
         int size = 16;
-        //gui setup code
-        //RenderSystem.clear(256, Minecraft.ON_OSX);
-        Matrix4f oldProjection = RenderSystem.getProjectionMatrix();
-        Matrix4f matrix4f = new Matrix4f().setOrtho(0.0F,
-                size, 0, size, 1000.0F, 3000); //ForgeHooksClient.getGuiFarPlane()
+        //save old projection and sets new orthographic
+        RenderSystem.backupProjectionMatrix();
+        //like this so object center is exactly at 0 0 0
+        Matrix4f matrix4f = new Matrix4f().setOrtho(0.0F, size, size, 0, -1000.0F, 1000);
         RenderSystem.setProjectionMatrix(matrix4f);
 
         //model view stuff
@@ -170,24 +161,13 @@ public class RenderedTexturesManager {
         posestack.pushPose();
         posestack.setIdentity();
 
-        posestack.translate(0.0D, 0.0D, -2000); //ForgeHooksClient.getGuiFarPlane()
         //apply new model view transformation
         RenderSystem.applyModelViewMatrix();
         Lighting.setupFor3DItems();
         //end gui setup code
 
-        drawFunction.accept(posestack);
-
-        var t = new Tesselator();
-        var b = t.getBuilder();
-        b.begin(VertexFormat.Mode.TRIANGLES, DefaultVertexFormat.POSITION_COLOR);
-
-        b.color(1.0f, 0.0f, 0.0f, 1);
-        b.vertex(0.0f, 0f,0);
-        b.vertex(5f, 5f, 0.0f);
-        b.vertex(0f, 5f, 0.0f);
-
-        BufferUploader.drawWithShader(b.end());
+        //item renderer needs a new pose stack as it applies its last to render system itself. for the rest tbh idk
+        drawFunction.accept(new PoseStack());
 
         //reset stuff
         posestack.popPose();
@@ -195,7 +175,7 @@ public class RenderedTexturesManager {
         RenderSystem.applyModelViewMatrix();
 
         //reset projection
-        RenderSystem.setProjectionMatrix(oldProjection);
+        RenderSystem.restoreProjectionMatrix();
         //RenderSystem.clear(256, Minecraft.ON_OSX);
         //returns render calls to main render target
         mc.getMainRenderTarget().bindWrite(true);