+++ fabric/src/main/java/net/mehvahdjukaar/moonlight/api/events/fabric/VillagerBrainEvent.java
@@ -6,6 +6,7 @@ import net.mehvahdjukaar.moonlight.api.events.IVillagerBrainEvent;
 import net.mehvahdjukaar.moonlight.core.misc.VillagerBrainEventInternal;
 import net.minecraft.world.entity.ai.Brain;
 import net.minecraft.world.entity.ai.behavior.Behavior;
+import net.minecraft.world.entity.ai.behavior.BehaviorControl;
 import net.minecraft.world.entity.ai.memory.ExpirableValue;
 import net.minecraft.world.entity.ai.memory.MemoryModuleType;
 import net.minecraft.world.entity.ai.sensing.Sensor;
@@ -58,7 +59,7 @@ public class VillagerBrainEvent implements IVillagerBrainEvent {
      * @param activity        the identifier of the activity
      * @param activityPackage the play package itself that will be executed
      */
-    public void addOrReplaceActivity(Activity activity, ImmutableList<? extends Pair<Integer, ? extends Behavior<? super Villager>>> activityPackage) {
+    public void addOrReplaceActivity(Activity activity, ImmutableList<? extends Pair<Integer, ? extends BehaviorControl<? super Villager>>> activityPackage) {
         this.internal.addOrReplaceActivity(activity, activityPackage);
     }