+++ changelog.md
@@ -1,3 +1,2 @@
-- fixed cave filter placement modifier
-- added to vanilla methods to wood type
-- fixed some item rendering methods
+- fixed texture renderer
+- fixed villager ai stuff